Move PubsubIO to separate project
