+++ scio-examples/src/main/scala/com/spotify/scio/examples/complete/game/LeaderBoard.scala
@@ -31,6 +31,7 @@ package com.spotify.scio.examples.complete.game
 import java.util.TimeZone
 
 import com.spotify.scio._
+import com.spotify.scio.pubsub._
 import com.spotify.scio.bigquery._
 import com.spotify.scio.examples.complete.game.UserScore.GameActionInfo
 import com.spotify.scio.streaming._
